@sitajaf & @francikas #95687036: worked on scripts image deplyoment

+++ scripts/deployment/install.sh
@@ -9,17 +9,19 @@ deployFolder=$1
 
 
 #Ensure that Docker is installed
-if [ ! -f /usr/local/bin/docker ]; then
-    #install docker
+if [ ! -f /usr/local/bin/docker ] || [ ! -f /usr/bin/sshpass ]; then
     apt-get update
-    apt-get -y install docker.io
+
+    if [ ! -f /usr/local/bin/docker ]; then
+    apt-get -y install docker.io;
     ln -sf /usr/bin/docker.io /usr/local/bin/docker
     sed -i '$acomplete -F _docker docker' /etc/bash_completion.d/docker.io
     update-rc.d docker.io defaults
+    fi
 
-    #install DOCKER-COMPOSE
-    curl -L https://github.com/docker/compose/releases/download/1.2.0/docker-compose-`uname -s`-`uname -m` > /usr/local/bin/docker-compose
-    chmod +x /usr/local/bin/docker-compose
+    if [ ! -f /usr/bin/sshpass ]; then
+    apt-get -y install sshpass
+    fi
 
 fi
 
@@ -27,3 +29,6 @@ mkdir -p /opt/app/eums
 
 #install the image
 scripts/install-image-eums.sh
+
+# uninstall ssh-pass
+apt-get -y --purge remove sshpass
